==> paper-analysis/191005-EvaluateModelLearningCurve/04-FinalModel.ipynb <==
# Train Final Model

import numpy as np
import pandas as pd
import joblib
# sklearn import
from sklearn.preprocessing import PolynomialFeatures, StandardScaler
from sklearn.linear_model import Lasso
from sklearn.pipeline import Pipeline
from sklearn.feature_selection import VarianceThreshold


# my module imports
from optimalcodon.projects.rnastability.dataprocessing import get_data, general_preprocesing_pipeline


(train_x, train_y), (test_x, test_y) = get_data('../data/191004-TrainAndTestSets/')
print("{} points for training and {} for testing with {} features".format(
    train_x.shape[0], test_x.shape[0], test_x.shape[1]))


# TODO: change for the new trained model
# we load this to get the params
lasso = joblib.load("results-data/best-model-gridserach.joblib")


model = Pipeline([
    ('general', general_preprocesing_pipeline(train_x)), # see the code for general_preprocesing_pipeline
    ('polyfeaturs', PolynomialFeatures(degree=2)),
    ('zerovar', VarianceThreshold(threshold=0.0)),
    ('scaling', StandardScaler()),
    ('lasso', Lasso(**lasso.get_params()))
])

model.fit(train_x, train_y)

## save trained model
joblib.dump(model, "results-data/final_model.joblib")

## predict test data
model.predict(test_x)

test_x['predicted'] = model.predict(test_x)
test_x['observed'] = test_y.values
(
    test_x
    .loc[:, ['specie', 'cell_type', 'datatype', 'predicted', 'observed']]
    .to_csv("results-data/test_data_predictions.csv")
)
